chore: refactor based on e2e code changes
