+++ package.json
@@ -5,7 +5,6 @@
   "repository": "github:argentlabs/argent-x",
   "homepage": "https://github.com/argentlabs/argent-x/#readme",
   "devDependencies": {
-    "@tsconfig/node18": "^1.0.1",
     "@types/prompt": "^1.1.2",
     "bundlewatch": "^0.3.3",
     "husky": "^8.0.0",
@@ -15,9 +14,7 @@
     "prettier": ">=2",
     "prettier-plugin-import-sort": "^0.0.7",
     "prompt": "^1.3.0",
-    "starknet": "^3.16.0",
-    "ts-node": "^10.8.1",
-    "typescript": "^4.7.4"
+    "ts-node": "^10.8.1"
   },
   "workspaces": [
     "packages/*"
@@ -33,9 +30,7 @@
     "test:e2e": "lerna run test:e2e --stream",
     "prepare": "yarn --cwd packages/get-starknet build && husky install",
     "storybook": "cd packages/storybook && yarn storybook",
-    "devnet:start": "docker run -it -p 5050:5050 shardlabs/starknet-devnet:latest --seed 0 --accounts 1",
-    "devnet:start:arm": "docker run -it -p 5050:5050 shardlabs/starknet-devnet:latest-arm --seed 0 --accounts 1",
-    "devnet:upgrade-helper": "NODE_NO_WARNINGS=1 ts-node ./scripts/devnet/devnet-upgrade-helper.ts"
+    "devnet:upgrade-helper": "NODE_NO_WARNINGS=1 ts-node ./scripts/devnet-upgrade-helper.ts"
   },
   "importSort": {
     ".js, .jsx, .ts, .tsx": {
