+++ packages/extension/e2e/apis/sendDevnetEthToAccount.ts
@@ -1,16 +1,6 @@
-import { Account, Provider, ec, number, stark, uint256 } from "starknet"
+import { number, stark, uint256 } from "starknet"
 
-// Spin up devnet with `starknet-devnet --seed 0` to get these values
-
-const pk = "0xe3e70682c2094cac629f6fbed82c07cd" // put hex representation of private key here
-
-const provider = new Provider({ baseUrl: "http://127.0.0.1:5050/" }) // connect to devnet
-const keyPair = ec.getKeyPair(pk)
-const account = new Account(
-  provider,
-  "0x7e00d496e324876bbc8531f2d9a82bf154d1a04a50218ee74cdd372f75a551a".toLowerCase(), // put account address which is controlled by the private key here
-  keyPair,
-)
+import { account } from "./account"
 
 export async function sendDevnetEthToAccount(address: string) {
   const response = await account.execute({
