+++ scripts/devnet-upgrade-helper.ts
@@ -1,0 +1,60 @@
+import prompt from "prompt"
+import { validateAndParseAddress } from "starknet"
+
+import { declareUpgradeContract } from "../packages/extension/e2e/apis/declareUpgradeContract"
+import { sendDevnetEthToAccount } from "../packages/extension/e2e/apis/sendDevnetEthToAccount"
+
+;(async () => {
+  /** clear default 'prompt' message */
+  prompt.message = ""
+
+  console.log("Declaring upgrade contract...")
+
+  const classHash = await declareUpgradeContract()
+
+  console.log(
+    `Default contract class hash: 0x3e327de1c40540b98d05cbcb13552008e36f0ec8d61d46956d2f9752c294328`,
+  )
+  console.log(`Upgrade contract class hash: ${classHash}`)
+  console.log("")
+
+  let previousAddress = ""
+  let keepAsking = true
+
+  while (keepAsking) {
+    console.log("Enter a localhost wallet address to transfer Ξ1")
+    previousAddress &&
+      previousAddress.length &&
+      console.log(`(Leave empty to use ${previousAddress}`)
+    try {
+      const result = await prompt.get([
+        {
+          name: "address",
+          description: "Address",
+          type: "string",
+        },
+      ])
+      const address: string = (result.address as string) || previousAddress
+      let addressIsValid
+      try {
+        addressIsValid = address && validateAndParseAddress(address)
+      } catch (e) {
+        // ignore validateAndParseAddress error
+      }
+      if (addressIsValid) {
+        previousAddress = address
+        console.log(`Transferring Ξ1 to ${address}`)
+        await sendDevnetEthToAccount(address)
+      } else {
+        console.log(`"${address}" is not valid wallet address`)
+      }
+    } catch (e) {
+      if (e instanceof Error && e.message === "canceled") {
+        // consume 'cancel' error from 'prompt', just cleanly stop
+        keepAsking = false
+      } else {
+        throw e
+      }
+    }
+  }
+})()
